@@ -45,43 +45,43 @@
   "private": true,
   "packageManager": "yarn@1.22.22+sha512.a6b2f7906b721bba3d67d4aff083df04dad64c399707841b7acf00f6b133b7ac24255f2652fa22ae3534329dc6180534e98d17432037ff6fd140556e2bb3137e",
   "dependencies": {
-    "@angular/animations": "^20.0.1",
-    "@angular/common": "^20.0.1",
-    "@angular/compiler": "^20.0.1",
-    "@angular/core": "^20.0.1",
-    "@angular/forms": "^20.0.1",
-    "@angular/platform-browser": "^20.0.1",
-    "@angular/platform-browser-dynamic": "^20.0.1",
-    "@angular/platform-server": "^20.0.1",
-    "@angular/router": "^20.0.1",
-    "@angular/ssr": "^20.0.1",
-    "rxjs": "7.8.1",
+    "@angular/animations": "^21.0.3",
+    "@angular/common": "^21.0.3",
+    "@angular/compiler": "^21.0.3",
+    "@angular/core": "^21.0.3",
+    "@angular/forms": "^21.0.3",
+    "@angular/platform-browser": "^21.0.3",
+    "@angular/platform-browser-dynamic": "^21.0.3",
+    "@angular/platform-server": "^21.0.3",
+    "@angular/router": "^21.0.3",
+    "@angular/ssr": "^21.0.2",
+    "rxjs": "7.8.2",
     "tslib": "^2.8.1",
     "zone.js": "~0.15.0"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "^0.2000.1",
-    "@angular-devkit/core": "^20.0.1",
-    "@angular-devkit/schematics": "^20.0.1",
-    "@angular-eslint/builder": "19.8.0",
-    "@angular-eslint/eslint-plugin": "19.8.0",
-    "@angular-eslint/eslint-plugin-template": "19.8.0",
-    "@angular-eslint/schematics": "19.8.0",
-    "@angular-eslint/template-parser": "19.8.0",
-    "@angular/build": "^20.0.1",
-    "@angular/cli": "^20.0.1",
-    "@angular/compiler-cli": "^20.0.1",
-    "@angular/localize": "20.0.1",
-    "@types/jasmine": "~5.1.4",
-    "@types/node": "^22.9.1",
-    "@typescript-eslint/eslint-plugin": "^8.15.0",
-    "@typescript-eslint/parser": "^8.15.0",
-    "@typescript-eslint/types": "^8.15.0",
-    "@typescript-eslint/utils": "^8.15.0",
-    "commit-and-tag-version": "^12.5.0",
-    "cross-env": "^7.0.3",
-    "eslint": "^8.57.1",
-    "eslint-plugin-import": "npm:eslint-plugin-i@^2.29.1",
+    "@angular-devkit/architect": "^0.2100.2",
+    "@angular-devkit/core": "^21.0.2",
+    "@angular-devkit/schematics": "^21.0.2",
+    "@angular-eslint/builder": "21.0.1",
+    "@angular-eslint/eslint-plugin": "21.0.1",
+    "@angular-eslint/eslint-plugin-template": "21.0.1",
+    "@angular-eslint/schematics": "21.0.1",
+    "@angular-eslint/template-parser": "21.0.1",
+    "@angular/build": "^21.0.2",
+    "@angular/cli": "^21.0.2",
+    "@angular/compiler-cli": "^21.0.3",
+    "@angular/localize": "21.0.3",
+    "@types/jasmine": "~5.1.13",
+    "@types/node": "^24.10.1",
+    "@typescript-eslint/eslint-plugin": "^8.33.1",
+    "@typescript-eslint/parser": "^8.33.1",
+    "@typescript-eslint/types": "^8.48.1",
+    "@typescript-eslint/utils": "^8.33.1",
+    "commit-and-tag-version": "^12.6.1",
+    "cross-env": "^10.1.0",
+    "eslint": "^9.28.0",
+    "eslint-plugin-import-x": "^4.16.1",
     "glob": "^10.4.5",
     "jasmine": "~5.4.0",
     "jasmine-core": "~5.4.0",
@@ -90,11 +90,11 @@
     "karma-coverage": "~2.2.1",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.1.0",
-    "ng-packagr": "^20.0.0",
+    "ng-packagr": "^21.0.0",
     "npm-run-all": "^4.1.5",
-    "prettier": "3.3.3",
-    "ts-node": "^10.9.2",
-    "typescript": "~5.8.3"
+    "prettier": "3.7.4",
+    "tsx": "^4.21.0",
+    "typescript": "~5.9.3"
   },
   "prettier": {
     "singleQuote": true,
