@@ -3,15 +3,16 @@
  * Copyright Google LLC All Rights Reserved.
  *
  * Use of this source code is governed by an MIT-style license that can be
- * found in the LICENSE file at https://angular.io/license
+ * found in the LICENSE file at https://angular.dev/license
  */
+
+/* eslint-disable import/no-extraneous-dependencies */
 
 import { Architect, BuilderOutput, ScheduleOptions, Target } from '@angular-devkit/architect';
 import { WorkspaceNodeModulesArchitectHost } from '@angular-devkit/architect/node';
 import { TestProjectHost, TestingArchitectHost } from '@angular-devkit/architect/testing';
 import {
   Path,
-  PathFragment,
   getSystemPath,
   join,
   json,
@@ -20,12 +21,14 @@ import {
   virtualFs,
   workspaces,
 } from '@angular-devkit/core';
+import path from 'node:path';
+import { firstValueFrom } from 'rxjs';
 
-// Default timeout for large specs is 2.5 minutes.
-jasmine.DEFAULT_TIMEOUT_INTERVAL = 150000;
+// Default timeout for large specs is 60s.
+jasmine.DEFAULT_TIMEOUT_INTERVAL = 60_000;
 
-export const applicationWorkspaceRoot = join(normalize(__dirname), `ng-application-app/`);
-export const applicationHost = new TestProjectHost(applicationWorkspaceRoot);
+export const workspaceRoot = join(normalize(__dirname), `../projects/hello-world-app/`);
+export const host = new TestProjectHost(workspaceRoot);
 export const outputPath: Path = normalize('dist');
 
 export const browserTargetSpec = { project: 'app', target: 'build' };
@@ -35,10 +38,17 @@ export const karmaTargetSpec = { project: 'app', target: 'test' };
 export const tslintTargetSpec = { project: 'app', target: 'lint' };
 export const protractorTargetSpec = { project: 'app-e2e', target: 'e2e' };
 
-export async function createArchitect(workspaceRoot: Path, host: TestProjectHost) {
+export async function createArchitect(workspaceRoot: Path) {
   const registry = new schema.CoreSchemaRegistry();
   registry.addPostTransform(schema.transforms.addUndefinedDefaults);
   const workspaceSysPath = getSystemPath(workspaceRoot);
+
+  // The download path is relative (set from Starlark), so before potentially
+  // changing directories, or executing inside a temporary directory, ensure
+  // the path is absolute.
+  if (process.env['PUPPETEER_DOWNLOAD_PATH']) {
+    process.env.PUPPETEER_DOWNLOAD_PATH = path.resolve(process.env['PUPPETEER_DOWNLOAD_PATH']);
+  }
 
   const { workspace } = await workspaces.readWorkspace(
     workspaceSysPath,
@@ -49,8 +59,6 @@ export async function createArchitect(workspaceRoot: Path, host: TestProjectHost
     workspaceSysPath,
     new WorkspaceNodeModulesArchitectHost(workspace, workspaceSysPath),
   );
-  await architectHost.addBuilderFromPackage('..');
-  await architectHost.addBuilderFromPackage('../../../../dist/angular-server-side-configuration');
   const architect = new Architect(architectHost, registry);
 
   return {
@@ -65,15 +73,6 @@ export interface BrowserBuildOutput {
   files: { [file: string]: Promise<string> };
 }
 
-const firstValueFrom = <T>(observable: any): Promise<T> =>
-  new Promise((resolve, reject) => {
-    observable.subscribe({
-      next: (value: any) => resolve(value),
-      error: (err: any) => reject(err),
-      complete: () => {},
-    });
-  });
-
 export async function browserBuild(
   architect: Architect,
   host: virtualFs.Host,
@@ -82,7 +81,7 @@ export async function browserBuild(
   scheduleOptions?: ScheduleOptions,
 ): Promise<BrowserBuildOutput> {
   const run = await architect.scheduleTarget(target, overrides, scheduleOptions);
-  const output = (await run.result) as BuilderOutput;
+  const output = (await run.result) as BuilderOutput & { outputs: { path: string }[] };
   expect(output.success).toBe(true);
 
   if (!output.success) {
@@ -94,11 +93,11 @@ export async function browserBuild(
     };
   }
 
-  const [{ path }] = output['outputs'];
+  const [{ path }] = output.outputs;
   expect(path).toBeTruthy();
   const outputPath = normalize(path);
 
-  const fileNames = await firstValueFrom<PathFragment[]>(host.list(outputPath));
+  const fileNames = await firstValueFrom(host.list(outputPath));
   const files = fileNames.reduce((acc: { [name: string]: Promise<string> }, path) => {
     let cache: Promise<string> | null = null;
     Object.defineProperty(acc, path, {
@@ -110,8 +109,8 @@ export async function browserBuild(
         if (!fileNames.includes(path)) {
           return Promise.reject('No file named ' + path);
         }
-        cache = firstValueFrom<virtualFs.FileBuffer>(host.read(join(outputPath, path))).then(
-          (content) => virtualFs.fileBufferToString(content),
+        cache = firstValueFrom(host.read(join(outputPath, path))).then((content) =>
+          virtualFs.fileBufferToString(content),
         );
 
         return cache;
@@ -128,3 +127,58 @@ export async function browserBuild(
     files,
   };
 }
+
+export const lazyModuleFiles: { [path: string]: string } = {
+  'src/app/lazy/lazy-routing.module.ts': `
+    import { NgModule } from '@angular/core';
+    import { Routes, RouterModule } from '@angular/router';
+
+    const routes: Routes = [];
+
+    @NgModule({
+      imports: [RouterModule.forChild(routes)],
+      exports: [RouterModule]
+    })
+    export class LazyRoutingModule { }
+  `,
+  'src/app/lazy/lazy.module.ts': `
+    import { NgModule } from '@angular/core';
+    import { CommonModule } from '@angular/common';
+
+    import { LazyRoutingModule } from './lazy-routing.module';
+
+    @NgModule({
+      imports: [
+        CommonModule,
+        LazyRoutingModule
+      ],
+      declarations: []
+    })
+    export class LazyModule { }
+  `,
+};
+
+export const lazyModuleFnImport: { [path: string]: string } = {
+  'src/app/app.module.ts': `
+    import { BrowserModule } from '@angular/platform-browser';
+    import { NgModule } from '@angular/core';
+
+    import { AppComponent } from './app.component';
+    import { RouterModule } from '@angular/router';
+
+    @NgModule({
+      declarations: [
+        AppComponent
+      ],
+      imports: [
+        BrowserModule,
+        RouterModule.forRoot([
+          { path: 'lazy', loadChildren: () => import('./lazy/lazy.module').then(m => m.LazyModule) }
+        ])
+      ],
+      providers: [],
+      bootstrap: [AppComponent]
+    })
+    export class AppModule { }
+`,
+};
